Conflicts:
	build.gradle
	build.properties
	src/main/java/li/cil/oc/api/API.java
	src/main/scala/li/cil/oc/client/Sound.scala
	src/main/scala/li/cil/oc/client/gui/DynamicGuiContainer.scala
	src/main/scala/li/cil/oc/common/EventHandler.scala
	src/main/scala/li/cil/oc/common/asm/ClassTransformer.scala
	src/main/scala/li/cil/oc/common/block/Adapter.scala
	src/main/scala/li/cil/oc/common/block/Assembler.scala
	src/main/scala/li/cil/oc/common/block/Capacitor.scala
	src/main/scala/li/cil/oc/common/block/Case.scala
	src/main/scala/li/cil/oc/common/block/Charger.scala
	src/main/scala/li/cil/oc/common/block/Disassembler.scala
	src/main/scala/li/cil/oc/common/block/DiskDrive.scala
	src/main/scala/li/cil/oc/common/block/Item.scala
	src/main/scala/li/cil/oc/common/block/Keyboard.scala
	src/main/scala/li/cil/oc/common/block/Microcontroller.scala
	src/main/scala/li/cil/oc/common/block/Raid.scala
	src/main/scala/li/cil/oc/common/block/RedstoneAware.scala
	src/main/scala/li/cil/oc/common/block/RobotAfterimage.scala
	src/main/scala/li/cil/oc/common/block/RobotProxy.scala
	src/main/scala/li/cil/oc/common/block/Screen.scala
	src/main/scala/li/cil/oc/common/block/ServerRack.scala
	src/main/scala/li/cil/oc/common/block/SimpleBlock.scala
	src/main/scala/li/cil/oc/common/block/Switch.scala
	src/main/scala/li/cil/oc/common/block/traits/SpecialBlock.scala
	src/main/scala/li/cil/oc/common/component/TextBuffer.scala
	src/main/scala/li/cil/oc/common/entity/Drone.scala
	src/main/scala/li/cil/oc/common/init/Items.scala
	src/main/scala/li/cil/oc/common/item/Tablet.scala
	src/main/scala/li/cil/oc/common/tileentity/Capacitor.scala
	src/main/scala/li/cil/oc/common/tileentity/Microcontroller.scala
	src/main/scala/li/cil/oc/common/tileentity/Redstone.scala
	src/main/scala/li/cil/oc/common/tileentity/Screen.scala
	src/main/scala/li/cil/oc/common/tileentity/ServerRack.scala
	src/main/scala/li/cil/oc/common/tileentity/traits/BundledRedstoneAware.scala
	src/main/scala/li/cil/oc/common/tileentity/traits/Computer.scala
	src/main/scala/li/cil/oc/common/tileentity/traits/RedstoneAware.scala
	src/main/scala/li/cil/oc/common/tileentity/traits/power/Mekanism.scala
	src/main/scala/li/cil/oc/integration/Mods.scala
	src/main/scala/li/cil/oc/integration/opencomputers/DriverBlockEnvironments.scala
	src/main/scala/li/cil/oc/server/component/Redstone.scala
	src/main/scala/li/cil/oc/server/component/RedstoneWireless.scala
	src/main/scala/li/cil/oc/server/component/UpgradeInventoryController.scala
	src/main/scala/li/cil/oc/server/component/robot/Player.scala
	src/main/scala/li/cil/oc/server/network/Network.scala
	src/main/scala/li/cil/oc/util/ItemCosts.scala
	src/main/scala/li/cil/oc/util/ItemUtils.scala

+++ src/main/scala/li/cil/oc/client/PacketSender.scala
@@ -76,35 +76,35 @@ object PacketSender {
     }
   }
 
-  def sendMouseClick(address: String, x: Int, y: Int, drag: Boolean, button: Int) {
+  def sendMouseClick(address: String, x: Double, y: Double, drag: Boolean, button: Int) {
     val pb = new SimplePacketBuilder(PacketType.MouseClickOrDrag)
 
     pb.writeUTF(address)
-    pb.writeShort(x)
-    pb.writeShort(y)
+    pb.writeFloat(x.toFloat)
+    pb.writeFloat(y.toFloat)
     pb.writeBoolean(drag)
     pb.writeByte(button.toByte)
 
     pb.sendToServer()
   }
 
-  def sendMouseScroll(address: String, x: Int, y: Int, scroll: Int) {
+  def sendMouseScroll(address: String, x: Double, y: Double, scroll: Int) {
     val pb = new SimplePacketBuilder(PacketType.MouseScroll)
 
     pb.writeUTF(address)
-    pb.writeShort(x)
-    pb.writeShort(y)
+    pb.writeFloat(x.toFloat)
+    pb.writeFloat(y.toFloat)
     pb.writeByte(scroll)
 
     pb.sendToServer()
   }
 
-  def sendMouseUp(address: String, x: Int, y: Int, button: Int) {
+  def sendMouseUp(address: String, x: Double, y: Double, button: Int) {
     val pb = new SimplePacketBuilder(PacketType.MouseUp)
 
     pb.writeUTF(address)
-    pb.writeShort(x)
-    pb.writeShort(y)
+    pb.writeFloat(x.toFloat)
+    pb.writeFloat(y.toFloat)
     pb.writeByte(button.toByte)
 
     pb.sendToServer()
